feat: Criação do model do sequelize
realizado a configuração e criação dos métodos create, findAll e findById do model

+++ backend/src/database/migrations/02-create-transactions.ts
@@ -1,6 +1,6 @@
 import { Model, QueryInterface, DataTypes } from 'sequelize'; 
 
-import { Transaction } from '../../types/Transaction';
+import { Transaction } from '../../interfaces/Transaction';
 
 export default {
   up(queryInterface: QueryInterface) {
@@ -26,7 +26,7 @@ export default {
 
       cashback: {
         allowNull: false,
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(10, 2),
       },
 
       amount: {
